requirements.txt
streamlit
pandas
scikit-learn
nltk
PyMuPDF==1.25.3
python-dotenv
plotly
openai
langchain
langchain-openai
langgraph
reportlab
joblib
pytesseract
pillow
